+++ backend/src/main/java/com/festago/auth/infrastructure/oauth2/KakaoOAuth2AccessTokenErrorHandler.java
@@ -1,4 +1,4 @@
-package com.festago.auth.infrastructure;
+package com.festago.auth.infrastructure.oauth2;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
