+++ backend/src/main/java/com/festago/auth/infrastructure/oauth2/KakaoOAuth2Client.java
@@ -1,4 +1,4 @@
-package com.festago.auth.infrastructure;
+package com.festago.auth.infrastructure.oauth2;
 
 import com.festago.auth.application.OAuth2Client;
 import com.festago.auth.domain.SocialType;
