+++ backend/src/main/java/com/festago/auth/infrastructure/openid/AppleOpenIdJwksClient.java
@@ -1,4 +1,4 @@
-package com.festago.auth.infrastructure;
+package com.festago.auth.infrastructure.openid;
 
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.InternalServerException;
@@ -12,18 +12,18 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.client.ResourceAccessException;
 import org.springframework.web.client.RestTemplate;
 
-@Slf4j
 @Component
-public class KakaoOpenIdJwksClient {
+@Slf4j
+public class AppleOpenIdJwksClient {
 
     private final RestTemplate restTemplate;
     private final Parser<JwkSet> parser;
 
-    public KakaoOpenIdJwksClient(
+    public AppleOpenIdJwksClient(
         RestTemplateBuilder restTemplateBuilder
     ) {
         this.restTemplate = restTemplateBuilder
-            .errorHandler(new KakaoOpenIdJwksErrorHandler())
+            .errorHandler(new AppleOpenIdJwksErrorHandler())
             .setConnectTimeout(Duration.ofSeconds(2))
             .setReadTimeout(Duration.ofSeconds(3))
             .build();
@@ -31,14 +31,13 @@ public class KakaoOpenIdJwksClient {
             .build();
     }
 
-    // 너무 많은 요청이 오면 차단될 수 있음
     public JwkSet requestGetJwks() {
         try {
-            String jsonKeys = restTemplate.getForObject("https://kauth.kakao.com/.well-known/jwks.json", String.class);
-            log.info("카카오 JWKS 공개키 목록을 조회했습니다.");
+            String jsonKeys = restTemplate.getForObject("https://appleid.apple.com/auth/keys", String.class);
+            log.info("Apple JWKS 공개키 목록을 조회했습니다.");
             return parser.parse(jsonKeys);
         } catch (ResourceAccessException e) {
-            log.warn("카카오 JWKS 서버가 응답하지 않습니다.");
+            log.warn("Apple JWKS 서버가 응답하지 않습니다.");
             throw new InternalServerException(ErrorCode.OPEN_ID_PROVIDER_NOT_RESPONSE);
         }
     }
