+++ backend/src/main/java/com/festago/auth/infrastructure/openid/KakaoOpenIdJwksErrorHandler.java
@@ -1,4 +1,4 @@
-package com.festago.auth.infrastructure;
+package com.festago.auth.infrastructure.openid;
 
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.InternalServerException;
