+++ backend/src/test/java/com/festago/auth/domain/OAuth2ClientsTest.java
@@ -8,10 +8,10 @@ import static org.mockito.Mockito.mock;
 import com.festago.auth.application.OAuth2Client;
 import com.festago.auth.application.OAuth2Clients;
 import com.festago.auth.application.OAuth2Clients.OAuth2ClientsBuilder;
-import com.festago.auth.infrastructure.FestagoOAuth2Client;
-import com.festago.auth.infrastructure.KakaoOAuth2AccessTokenClient;
-import com.festago.auth.infrastructure.KakaoOAuth2Client;
-import com.festago.auth.infrastructure.KakaoOAuth2UserInfoClient;
+import com.festago.auth.infrastructure.oauth2.FestagoOAuth2Client;
+import com.festago.auth.infrastructure.oauth2.KakaoOAuth2AccessTokenClient;
+import com.festago.auth.infrastructure.oauth2.KakaoOAuth2Client;
+import com.festago.auth.infrastructure.oauth2.KakaoOAuth2UserInfoClient;
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.UnexpectedException;
 import org.junit.jupiter.api.DisplayNameGeneration;
